+++ src/main/java/com/depromeet/stonebed/domain/fcm/application/FcmService.java
@@ -1,18 +1,11 @@
 package com.depromeet.stonebed.domain.fcm.application;
 
-import com.depromeet.stonebed.domain.fcm.dao.FcmRepository;
 import com.depromeet.stonebed.domain.fcm.domain.FcmResponseErrorType;
-import com.depromeet.stonebed.domain.fcm.domain.FcmToken;
-import com.depromeet.stonebed.domain.member.domain.Member;
-import com.depromeet.stonebed.global.error.ErrorCode;
-import com.depromeet.stonebed.global.error.exception.CustomException;
-import com.depromeet.stonebed.global.util.MemberUtil;
 import com.google.firebase.messaging.*;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.List;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -24,12 +17,11 @@ import org.springframework.transaction.annotation.Transactional;
 public class FcmService {
     private static final int BATCH_SIZE = 500;
 
-    private final FcmRepository fcmRepository;
-    private final MemberUtil memberUtil;
+    private final FcmTokenService fcmTokenService;
 
     @Transactional(readOnly = true)
     public void sendMulticastMessageToAll(Notification notification) {
-        List<String> tokens = getAllTokens();
+        List<String> tokens = fcmTokenService.getAllTokens();
         int totalTokens = tokens.size();
 
         for (int i = 0; i < totalTokens; i += BATCH_SIZE) {
@@ -83,7 +75,7 @@ public class FcmService {
                             String token =
                                     tokens.get(response.getResponses().indexOf(sendResponse));
                             if (isInvalidOrNotRegistered(sendResponse)) {
-                                invalidateToken(token);
+                                fcmTokenService.invalidateToken(token);
                             }
                         });
     }
@@ -93,56 +85,5 @@ public class FcmService {
         return FcmResponseErrorType.contains(errorMessage, FcmResponseErrorType.NOT_REGISTERED)
                 || FcmResponseErrorType.contains(
                         errorMessage, FcmResponseErrorType.INVALID_REGISTRATION);
-    }
-
-    @Transactional
-    public void invalidateToken(String token) {
-        fcmRepository
-                .findByToken(token)
-                .ifPresentOrElse(
-                        fcmToken -> updateToken(fcmToken, ""),
-                        () -> {
-                            throw new CustomException(ErrorCode.FAILED_TO_FIND_FCM_TOKEN);
-                        });
-    }
-
-    private void updateToken(FcmToken fcmToken, String token) {
-        fcmToken.updateToken(token);
-        fcmRepository.save(fcmToken);
-    }
-
-    @Transactional(readOnly = true)
-    public List<String> getAllTokens() {
-        return fcmRepository.findAll().stream()
-                .map(FcmToken::getToken)
-                .filter(token -> !token.isEmpty())
-                .toList();
-    }
-
-    @Transactional
-    public void storeOrUpdateToken(String token) {
-        final Member member = memberUtil.getCurrentMember();
-        Optional<FcmToken> existingToken = fcmRepository.findByMember(member);
-        existingToken.ifPresentOrElse(
-                fcmToken -> {
-                    fcmToken.updateToken(token);
-                    fcmRepository.save(fcmToken);
-                },
-                () -> {
-                    FcmToken fcmToken = new FcmToken(member, token);
-                    fcmRepository.save(fcmToken);
-                });
-    }
-
-    @Transactional
-    public void refreshTokenTimestampForCurrentUser() {
-        Member member = memberUtil.getCurrentMember();
-        fcmRepository
-                .findByMember(member)
-                .ifPresentOrElse(
-                        fcmToken -> updateToken(fcmToken, fcmToken.getToken()),
-                        () -> {
-                            throw new CustomException(ErrorCode.FAILED_TO_FIND_FCM_TOKEN);
-                        });
     }
 }
